Update to take care about logged user

+++ mint/app/Http/Controllers/MentorallconnectionController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use App\Collaboration;
 use App\User;
 
@@ -13,10 +14,15 @@ class MentorallconnectionController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index($id)
+    public function index()
     {
-        $menteeRequests = Collaboration::where('status_rqs', 'connected')->where('mentor_id', $id)->get();
-        return view('mentorAllConnection', ['menteeRequests' => $menteeRequests]);
+        if(Auth::check()){
+            $id = Auth::user()->id;
+            $menteeRequests = Collaboration::where('status_rqs', 'connected')->where('mentor_id', $id)->get();
+            return view('mentorAllConnection', ['menteeRequests' => $menteeRequests]);
+        }else{
+            return redirect('/');
+        }
     }
 
     /**
@@ -82,12 +88,16 @@ class MentorallconnectionController extends Controller
      */
     public function destroy($id)
     {
-        $deleteConn = Collaboration::find($id);
-        $deleteConn->delete();
-        if ($deleteConn) {
-            return response()->json(['msg'=>"Connection removed for $id"]);
+        if(Auth::check()){
+            $deleteConn = Collaboration::find($id);
+            $deleteConn->delete();
+            if ($deleteConn) {
+                return response()->json(['msg'=>"Connection removed for $id"]);
+            }else{
+                return response()->json(['msg'=>'Something wrong happened, collaboration not deleted']);
+            }
         }else{
-            return response()->json(['msg'=>'Something wrong happened, collaboration not deleted']);
+            return redirect('/');
         }
     }
 }
