+++ mint/app/Http/Controllers/MentorallinvitationController.php
@@ -16,9 +16,13 @@ class MentorallinvitationController extends Controller
      */
     public function index()
     {
-        $id = Auth::user()->id;
-        $menteeRequests = Collaboration::where('status_rqs', 'pending')->where('mentor_id', $id)->get();
-        return view('mentorAllInvitation', ['menteeRequests' => $menteeRequests]);
+        if(Auth::check()){
+            $id = Auth::user()->id;
+            $menteeRequests = Collaboration::where('status_rqs', 'pending')->where('mentor_id', $id)->get();
+            return view('mentorAllInvitation', ['menteeRequests' => $menteeRequests]);
+        }else{
+            return redirect('/');
+        }
     }
 
     /**
@@ -73,11 +77,15 @@ class MentorallinvitationController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $acceptCollab = Collaboration::find($id);
-        $acceptCollab->status_rqs = "connected";
-        // return response()->json($acceptCollab);
-        $acceptCollab->save();
-        return response()->json(['msg'=>"Invitation accepted for $id"]);
+        if(Auth::check()){
+            $acceptCollab = Collaboration::find($id);
+            $acceptCollab->status_rqs = "connected";
+            // return response()->json($acceptCollab);
+            $acceptCollab->save();
+            return response()->json(['msg'=>"Invitation accepted for $id"]);
+        }else{
+            return redirect('/');
+        }
     }
 
     /**
@@ -88,12 +96,16 @@ class MentorallinvitationController extends Controller
      */
     public function destroy($id)
     {
-        $declineCollab = Collaboration::find($id);
-        $declineCollab->delete();
-        if ($declineCollab) {
-            return response()->json(['msg'=>"Invitation decline for $id"]);
+        if(Auth::check()){
+            $declineCollab = Collaboration::find($id);
+            $declineCollab->delete();
+            if ($declineCollab) {
+                return response()->json(['msg'=>"Invitation decline for $id"]);
+            }else{
+                return response()->json(['msg'=>'Something wrong happened, Invitation decline not worked']);
+            }
         }else{
-            return response()->json(['msg'=>'Something wrong happened, Invitation decline not worked']);
+            return redirect('/');
         }
     }
 }
